requirements.txt
selenium==4.15.2
python-telegram-bot==20.7
APScheduler==3.10.4
python-dotenv==1.0.0
requests==2.31.0
webdriver-manager==4.0.1
